started my portfolio website, added all the images and functionalities needed.
torn between keeping the 'Home'  text on the home page or remove it to reduce redundancy. the thing is i don't know what to include in my home page anymore. anw, for my next step, i'll gather the projects i've worked on the past and include it in the website.